Implement tests for from_url with user/password
